requirements.txt
python-dotenv==0.7.1
Flask==1.0.2
Flask-Dance==1.0.0
gunicorn==19.8.1
